ble_scanner/requirements.txt
Flask==2.3.3
paho-mqtt==1.6.1
requests==2.31.0 